Prevent GitHub from normalizing asset filenames
Our pacman packages can contain "~" in their name if they have an epoch set in
their version. GitHub will silently replace that with "." on upload.

To prevent this we use the base64 encoded version of the basename as the asset
name and set the label, which should remain unchanged and which is visible as the asset
name in the GH UI, to the real filename.

Fixes #15