fetch-assets: be more strict in what we allow to download
In case any builds are in the queue that depend on finished builds in the
queue don't download both.

In case finished builds are in the queue and new dependencies get added
to the queue don't download both.

Handle this by grouping into separate dependency graphs and only download
finished sub graphs.

Might need a bit more feedback as to what and why it is skipped.

Fixes #19