Make msys2 config/installs idempotent.
To guard against the case that we may be reusing an msys2 install.

Also, increase job timeout.  According to Github docs, jobs on runners
they host are limited to 6 hours (which is also the default job limit),
but self-hosted runners can run up to the maximum runtime for a
workflow, 72 hours (4320 minutes)