Add a list of optional dependencies
Since we no longer break cycles in msys2-web we have to do it here.
This adds a list of optional deps for some packages. If they are there
they will be used, if not they will be ignored.

By hardcoding it we should get more a more deterministic result, but
not sure if this scales well.